requirements.txt
robotframework
robotframework-jsonlibrary
robotframework-pythonlibcore
robotframework-requests
robotframework-seleniumlibrary
robotframework-jsonschemalibrary